move reused context manager to a common location
